fix: refactor widget addition sagas and widget operation sagas to allow gridPosition in pasteSaga (#32992)
## Description
> [!TIP]  
**Problem**
The current building block on canvas implementation uses the pasteSaga
logic, and this logic requires the mousePosition to know where to paste
the widgets. Due to building blocks being mimicked as a single widget,
we get grid position values instead of mouse location when dragging and
dropping building blocks, this forced us to create e utility function
that generates mousePosition from gridPosition values. This logic was
flawed and led to building blocks placement being wrong when the view is
scrolled.
**Solution**
We have decided to implement the pasteSaga to allow both mousePosition
and gridPosition as expected values for dropping the copied widgets unto
the canvas.


Fixes #32962

## Automation

/ok-to-test tags="@tag.Widget"

### :mag: Cypress test results
