+++ app/client/src/sagas/WidgetAdditionSagas.ts
@@ -1,8 +1,3 @@
-import type {
-  ImportBuildingBlockToApplicationRequest,
-  ImportBuildingBlockToApplicationResponse,
-} from "@appsmith/api/ApplicationApi";
-import ApplicationApi from "@appsmith/api/ApplicationApi";
 import type { ReduxAction } from "@appsmith/constants/ReduxActionConstants";
 import {
   ReduxActionErrorTypes,
@@ -10,12 +5,6 @@ import {
   WidgetReduxActionTypes,
 } from "@appsmith/constants/ReduxActionConstants";
 import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
-import type { ActionDataState } from "@appsmith/reducers/entityReducers/actionsReducer";
-import {
-  getActions,
-  getCanvasWidgets,
-} from "@appsmith/selectors/entitiesSelector";
-import { getCurrentWorkspaceId } from "@appsmith/selectors/selectedWorkspaceSelectors";
 import type { WidgetBlueprint } from "WidgetProvider/constants";
 import {
   BlueprintOperationTypes,
@@ -25,14 +14,8 @@ import WidgetFactory from "WidgetProvider/factory";
 import { generateAutoHeightLayoutTreeAction } from "actions/autoHeightActions";
 import type { WidgetAddChild } from "actions/pageActions";
 import { updateAndSaveLayout } from "actions/pageActions";
-import { runAction } from "actions/pluginActionActions";
-import { pasteWidget } from "actions/widgetActions";
-import { selectWidgetInitAction } from "actions/widgetSelectionActions";
-import type { ApiResponse } from "api/ApiResponses";
-import type { Template } from "api/TemplatesApi";
 import {
   BUILDING_BLOCK_EXPLORER_TYPE,
-  MAIN_CONTAINER_WIDGET_ID,
   RenderModes,
 } from "constants/WidgetConstants";
 import { toast } from "design-system";
@@ -48,27 +31,19 @@ import type {
   CanvasWidgetsReduxState,
   FlattenedWidgetProps,
 } from "reducers/entityReducers/canvasWidgetsReducer";
-import type { DragDetails } from "reducers/uiReducers/dragResizeReducer";
-import { all, call, put, select, take, takeEvery } from "redux-saga/effects";
+import { all, call, put, select, takeEvery } from "redux-saga/effects";
 import { getDataTree } from "selectors/dataTreeSelectors";
 import {
   getCanvasWidth,
-  getCurrentApplicationId,
-  getCurrentPageId,
   getIsAutoLayout,
   getIsAutoLayoutMobileBreakPoint,
 } from "selectors/editorSelectors";
-import { getTemplatesSelector } from "selectors/templatesSelectors";
 import AppsmithConsole from "utils/AppsmithConsole";
 import { getNextEntityName } from "utils/AppsmithUtils";
 import { generateWidgetProps } from "utils/WidgetPropsUtils";
 import { generateReactKey } from "utils/generators";
-import { getCopiedWidgets, saveCopiedWidgets } from "utils/storage";
 import type { WidgetProps } from "widgets/BaseWidget";
 import { isStack } from "../layoutSystems/autolayout/utils/AutoLayoutUtils";
-import { saveBuildingBlockWidgetsToStore } from "./BuildingBlocksSagas";
-import { validateResponse } from "./ErrorSagas";
-import { postPageAdditionSaga } from "./TemplatesSagas";
 import {
   buildWidgetBlueprint,
   executeWidgetBlueprintBeforeOperations,
@@ -76,22 +51,8 @@ import {
   traverseTreeAndExecuteBlueprintChildOperations,
 } from "./WidgetBlueprintSagas";
 import { getPropertiesToUpdate } from "./WidgetOperationSagas";
-import {
-  getDefaultCanvas,
-  getMousePositionFromCanvasGridPosition,
-  getSnappedGrid,
-} from "./WidgetOperationUtils";
-import { SelectionRequestType } from "./WidgetSelectUtils";
-import {
-  getDragDetails,
-  getWidget,
-  getWidgetByName,
-  getWidgets,
-} from "./selectors";
-
-import { getBuildingBlockDragStartTimestamp } from "selectors/buildingBlocksSelectors";
-import { initiateBuildingBlockDropEvent } from "utils/buildingBlockUtils";
-import AnalyticsUtil from "@appsmith/utils/AnalyticsUtil";
+import { getWidget, getWidgets } from "./selectors";
+import { addBuildingBlockToCanvasSaga } from "./BuildingBlockAdditionSagas";
 
 const WidgetTypes = WidgetFactory.widgetTypes;
 
@@ -528,245 +489,13 @@ function* addNewTabChildSaga(
   yield put(updateAndSaveLayout(updatedWidgets));
 }
 
-function* addBuildingBlockActionsToApp(dragDetails: DragDetails) {
-  const applicationId: string = yield select(getCurrentApplicationId);
-  const buildingblockName = dragDetails.newWidget.displayName;
-  const buildingBlocks: Template[] = yield select(getTemplatesSelector);
-  const currentPageId: string = yield select(getCurrentPageId);
-  const workspaceId: string = yield select(getCurrentWorkspaceId);
-  const selectedBuildingBlock = buildingBlocks.find(
-    (buildingBlock) => buildingBlock.title === buildingblockName,
-  ) as Template;
-
-  const body: ImportBuildingBlockToApplicationRequest = {
-    pageId: currentPageId,
-    applicationId,
-    workspaceId,
-    templateId: selectedBuildingBlock.id,
-  };
-
-  // api call adds DS, queries and JS to page and returns new page dsl with building block
-  const response: ApiResponse<ImportBuildingBlockToApplicationResponse> =
-    yield call(ApplicationApi.importBuildingBlockToApplication, body);
-
-  return response;
-}
-
-function* getBuildingBlocksDropMousePosition(
-  topRow: number,
-  leftColumn: number,
-) {
-  const canvasWidgets: CanvasWidgetsReduxState = yield select(getCanvasWidgets);
-  let mousePosition = { x: 0, y: 0 };
-
-  // convert grid position to mouse position for paste functionality
-  const { canvasDOM, canvasId, containerWidget } =
-    getDefaultCanvas(canvasWidgets);
-  if (!canvasDOM || !containerWidget || !canvasId) {
-    mousePosition = { x: 0, y: 0 };
-  } else {
-    const canvasRect = canvasDOM.getBoundingClientRect();
-    const { padding, snapGrid } = getSnappedGrid(
-      containerWidget,
-      canvasRect.width,
-    );
-    mousePosition = getMousePositionFromCanvasGridPosition(
-      topRow,
-      leftColumn,
-      snapGrid,
-      padding,
-      canvasId as string,
-    );
-  }
-
-  return mousePosition;
-}
-
-function* runSingleAction(actionId: string) {
-  yield put(runAction(actionId));
-  yield take(ReduxActionTypes.RUN_ACTION_SUCCESS);
-}
-
-function* runNewlyCreatedActions(
-  actionsBeforeAddingBuildingBlock: ActionDataState,
-  actionsAfterAddingBuildingBlocks: ActionDataState,
-) {
-  const actionIdsBeforeAddingBB = actionsBeforeAddingBuildingBlock.map(
-    (obj) => obj.config.id,
-  );
-
-  const newlyAddedActions = actionsAfterAddingBuildingBlocks.filter(
-    (obj) => !actionIdsBeforeAddingBB.includes(obj.config.id),
-  );
-
-  // Run each action sequentially. We have a max of 2-3 actions per building block.
-  // If we run this in parallel, we will have a racing condition when multiple building blocks are drag and dropped quickly.
-  for (const action of newlyAddedActions) {
-    if (action.config.executeOnLoad) {
-      yield runSingleAction(action.config.id);
-    }
-  }
-}
-
-export function* addBuildingBlockToApplication(
-  buildingBlockWidget: WidgetAddChild,
-  skeletonLoaderId: string,
-) {
-  const { leftColumn, topRow } = buildingBlockWidget;
-  try {
-    const dragDetails: DragDetails = yield select(getDragDetails);
-    const applicationId: string = yield select(getCurrentApplicationId);
-    const workspaceId: string = yield select(getCurrentWorkspaceId);
-    const actionsBeforeAddingBuildingBlock: ActionDataState =
-      yield select(getActions);
-    const existingCopiedWidgets: unknown = yield call(getCopiedWidgets);
-    const buildingBlockDragStartTimestamp: number = yield select(
-      getBuildingBlockDragStartTimestamp,
-    );
-
-    // start loading for dragging building blocks
-    yield put({
-      type: ReduxActionTypes.DRAGGING_BUILDING_BLOCK_TO_CANVAS_INIT,
-    });
-
-    // makes sure updateAndSaveLayout completes first for skeletonWidget addition
-    yield take(ReduxActionTypes.SAVE_PAGE_SUCCESS);
-
-    const response: ApiResponse<ImportBuildingBlockToApplicationResponse> =
-      yield call(addBuildingBlockActionsToApp, dragDetails);
-    const isValid: boolean = yield validateResponse(response);
-
-    if (isValid) {
-      yield saveBuildingBlockWidgetsToStore(response);
-
-      const mousePosition: { x: number; y: number } = yield call(
-        getBuildingBlocksDropMousePosition,
-        topRow,
-        leftColumn,
-      );
-
-      // remove skeleton loader just before pasting the building block
-      yield put({
-        type: WidgetReduxActionTypes.WIDGET_SINGLE_DELETE,
-        payload: {
-          widgetId: skeletonLoaderId,
-          parentId: MAIN_CONTAINER_WIDGET_ID,
-          disallowUndo: true,
-          isShortcut: false,
-        },
-      });
-
-      yield put(pasteWidget(false, mousePosition));
-      const timeTakenToDropWidgetsInSeconds =
-        (Date.now() - buildingBlockDragStartTimestamp) / 1000;
-      yield call(postPageAdditionSaga, applicationId);
-      // remove selecting of recently pasted widgets caused by pasteWidget
-      yield put(selectWidgetInitAction(SelectionRequestType.Empty));
-
-      // stop loading after pasting process is complete
-      yield put({
-        type: ReduxActionTypes.DRAGGING_BUILDING_BLOCK_TO_CANVAS_SUCCESS,
-      });
-
-      const actionsAfterAddingBuildingBlocks: ActionDataState =
-        yield select(getActions);
-
-      if (
-        response.data.onPageLoadActions &&
-        response.data.onPageLoadActions.length > 0
-      ) {
-        yield runNewlyCreatedActions(
-          actionsBeforeAddingBuildingBlock,
-          actionsAfterAddingBuildingBlocks,
-        );
-      }
-
-      const timeTakenToCompleteInMs = buildingBlockDragStartTimestamp
-        ? Date.now() - buildingBlockDragStartTimestamp
-        : 0;
-      const timeTakenToCompleteInSeconds = timeTakenToCompleteInMs / 1000;
-
-      AnalyticsUtil.logEvent("DROP_BUILDING_BLOCK_COMPLETED", {
-        applicationId,
-        workspaceId,
-        source: "explorer",
-        eventData: {
-          buildingBlockName: dragDetails.newWidget.displayName,
-          timeTakenToCompletion: timeTakenToCompleteInSeconds,
-          timeTakenToDropWidgets: timeTakenToDropWidgetsInSeconds,
-        },
-      });
-      yield put({
-        type: ReduxActionTypes.RESET_BUILDING_BLOCK_DRAG_START_TIME,
-      });
-
-      if (existingCopiedWidgets) {
-        yield call(saveCopiedWidgets, JSON.stringify(existingCopiedWidgets));
-      }
-    }
-  } catch (error) {
-    yield put({
-      type: WidgetReduxActionTypes.WIDGET_SINGLE_DELETE,
-      payload: {
-        widgetId: skeletonLoaderId,
-        parentId: MAIN_CONTAINER_WIDGET_ID,
-        disallowUndo: true,
-        isShortcut: false,
-      },
-    });
-    yield put({
-      type: ReduxActionErrorTypes.DRAGGING_BUILDING_BLOCK_TO_CANVAS_ERROR,
-    });
-  }
-}
-
-function* addBuildingBlockSaga(addEntityAction: ReduxAction<WidgetAddChild>) {
-  const applicationId: string = yield select(getCurrentApplicationId);
-  const workspaceId: string = yield select(getCurrentWorkspaceId);
-  const dragDetails: DragDetails = yield select(getDragDetails);
-  const buildingblockName = dragDetails.newWidget.displayName;
-  const skeletonWidgetName = `loading_${buildingblockName
-    .toLowerCase()
-    .replace(/ /g, "_")}`;
-  const addSkeletonWidgetAction: ReduxAction<
-    WidgetAddChild & { shouldReplay: boolean }
-  > = {
-    ...addEntityAction,
-    payload: {
-      ...addEntityAction.payload,
-      type: "SKELETON_WIDGET",
-      widgetName: skeletonWidgetName,
-      widgetId: MAIN_CONTAINER_WIDGET_ID,
-      // so that the skeleton loader does not get included when the users uses the undo/redo
-      shouldReplay: false,
-    },
-  };
-
-  yield call(initiateBuildingBlockDropEvent, {
-    applicationId,
-    workspaceId,
-    buildingblockName,
-  });
-
-  yield call(addChildSaga, addSkeletonWidgetAction);
-  const skeletonWidget: FlattenedWidgetProps = yield select(
-    getWidgetByName,
-    skeletonWidgetName,
-  );
-  yield call(
-    addBuildingBlockToApplication,
-    addEntityAction.payload,
-    skeletonWidget.widgetId,
-  );
-}
-
 function* addUIEntitySaga(addEntityAction: ReduxAction<WidgetAddChild>) {
   try {
     const { payload } = addEntityAction;
     const { type } = payload;
 
     if (type === BUILDING_BLOCK_EXPLORER_TYPE) {
-      yield call(addBuildingBlockSaga, addEntityAction);
+      yield call(addBuildingBlockToCanvasSaga, addEntityAction);
     } else {
       yield call(addChildSaga, addEntityAction);
     }
