+++ app/client/src/sagas/WidgetOperationSagas.tsx
@@ -9,6 +9,10 @@ import {
 } from "@appsmith/constants/ReduxActionConstants";
 import { resetWidgetMetaProperty } from "actions/metaActions";
 import { selectWidgetInitAction } from "actions/widgetSelectionActions";
+import type {
+  EitherMouseLocationORGridPosition,
+  PasteWidgetReduxAction,
+} from "constants/WidgetConstants";
 import {
   GridDefaults,
   MAIN_CONTAINER_WIDGET_ID,
@@ -1162,14 +1166,15 @@ export function calculateNewWidgetPosition(
  * @param copiedTotalWidth total width of the copied widgets
  * @param copiedTopMostRow top row of the top most copied widget
  * @param copiedLeftMostColumn left column of the left most copied widget
+ * @param gridPosition left and top canvas grid position values
  * @returns
  */
 const getNewPositions = function* (
   copiedWidgetGroups: CopiedWidgetGroup[],
-  mouseLocation: { x: number; y: number },
   copiedTotalWidth: number,
   copiedTopMostRow: number,
   copiedLeftMostColumn: number,
+  whereToPasteWidget: EitherMouseLocationORGridPosition,
 ) {
   const selectedWidgetIDs: string[] = yield select(getSelectedWidgets);
   const canvasWidgets: CanvasWidgetsReduxState = yield select(getWidgets);
@@ -1215,12 +1220,12 @@ const getNewPositions = function* (
   const newPastingPositionDetails: NewPastePositionVariables = yield call(
     getNewPositionsBasedOnMousePositions,
     copiedWidgetGroups,
-    mouseLocation,
     selectedWidgets,
     canvasWidgets,
     copiedTotalWidth,
     copiedTopMostRow,
     copiedLeftMostColumn,
+    whereToPasteWidget,
   );
   return newPastingPositionDetails;
 };
@@ -1364,16 +1369,17 @@ function* getNewPositionsBasedOnSelectedWidgets(
  * @param copiedTotalWidth total width of the copied widgets
  * @param copiedTopMostRow top row of the top most copied widget
  * @param copiedLeftMostColumn left column of the left most copied widget
+ * @param gridPosition left and top canvas grid position values
  * @returns
  */
 function* getNewPositionsBasedOnMousePositions(
   copiedWidgetGroups: CopiedWidgetGroup[],
-  mouseLocation: { x: number; y: number },
   selectedWidgets: WidgetProps[],
   canvasWidgets: CanvasWidgetsReduxState,
   copiedTotalWidth: number,
   copiedTopMostRow: number,
   copiedLeftMostColumn: number,
+  whereToPasteWidget: EitherMouseLocationORGridPosition,
 ) {
   let { canvasDOM, canvasId, containerWidget } =
     getDefaultCanvas(canvasWidgets);
@@ -1395,13 +1401,15 @@ function* getNewPositionsBasedOnMousePositions(
   );
 
   // get mouse positions in terms of grid rows and columns of the pasting canvas
-  const mousePositions = getMousePositions(
-    canvasRect,
-    canvasId,
-    snapGrid,
-    padding,
-    mouseLocation,
-  );
+  const mousePositions = whereToPasteWidget.gridPosition
+    ? whereToPasteWidget.gridPosition
+    : getMousePositions(
+        canvasRect,
+        canvasId,
+        snapGrid,
+        padding,
+        whereToPasteWidget.mouseLocation,
+      );
 
   if (!snapGrid || !mousePositions) return {};
 
@@ -1483,14 +1491,11 @@ function* getNewPositionsBasedOnMousePositions(
 }
 
 /**
- * this saga create a new widget from the copied one to store
+ * This saga create a new widget from the copied one to store.
+ * It allows using both mouseLocation or gridPosition to locate where the copied widgets should be dropped.
+ * If gridPosition is available, use it, else, calculate gridPosition from mousePosition
  */
-function* pasteWidgetSaga(
-  action: ReduxAction<{
-    groupWidgets: boolean;
-    mouseLocation: { x: number; y: number };
-  }>,
-) {
+function* pasteWidgetSaga(action: ReduxAction<PasteWidgetReduxAction>) {
   const {
     flexLayers,
     widgets: copiedWidgets,
@@ -1580,10 +1585,10 @@ function* pasteWidgetSaga(
         yield call(
           getNewPositions,
           copiedWidgetGroups,
-          action.payload.mouseLocation,
           copiedTotalWidth,
           topMostWidget.topRow,
           leftMostWidget.leftColumn,
+          action.payload,
         ));
 
       if (canvasId) pastingIntoWidgetId = canvasId;
